+++ src/adapters/DefaultAdapter.ts
@@ -1,12 +1,12 @@
 export default class DefaultAdapter {
-  build(Model: any, props: Record<string, any>): any {
+  build<M = any>(Model: any, props: Record<string, any>): M {
     return new Model(props);
   }
-  async save(model: any, Model: any): Promise<any> {
-    return Promise.resolve(model.save()).then(() => model);
+  async save<M = any>(model: M, Model: any): Promise<M> {
+    return Promise.resolve((model as any).save()).then(() => model);
   }
-  async destroy(model: any, Model: any): Promise<any> {
-    return Promise.resolve(model.destroy()).then(() => model);
+  async destroy<M = any>(model: M, Model: any): Promise<M> {
+    return Promise.resolve((model as any).destroy()).then(() => model);
   }
   get(model: any, attr: string, Model?: any): any {
     return model.get(attr);